create gcp_keys in folder of script
the README suggests running "anthos/create_service_accounts.sh".   when the user does so, "." is the top-level dir.  since the script creates the keys in the dir ./gcp_keys, this puts gcp_keys in the top-level directory rather than in the anthos/ subdir as would happen in the manual method.

this is inconsistent.

with my proposed change, the gcp_keys dir will always be created in the parent dir of the script (anthos/), rather than the current directory (which would be the top-level dir if the user is following the directions in the README).